+++ CHANGELOG.rst
@@ -4,6 +4,7 @@ Changelog
 master
 ------
 
+- (`#186 <https://github.com/openscm/scmdata/pull/186>`_ and `#187 <https://github.com/openscm/scmdata/pull/187>`_) Fix the handling of non-alphanumeric characters in filenames on Windows for :class:`scmdata.database.ScmDatabase`. ``*`` values are no longer included in :class:`scmdata.database.ScmDatabase` filenames
 - (`#185 <https://github.com/openscm/scmdata/pull/185>`_) Allow :class:`scmdata.run.ScmRun` to read remote files by providing a URL to the constructor
 - (`#183 <https://github.com/openscm/scmdata/pull/183>`_) Deprecate :func:`scmdata.ops.integrate`, replacing with to :func:`scmdata.ops.cumsum` and :func:`scmdata.ops.cumtrapz`
 - (`#184 <https://github.com/openscm/scmdata/pull/184>`_) Add :func:`scmdata.run.ScmRun.round`